test: parameters should be a float

+++ tests/ghmm/test_algorithm.py
@@ -8,11 +8,11 @@ import ghmm.algorithm as algo
 class test_Alogirthm(unittest.TestCase):
     tolerance = 1e-10
 
-    O = np.array(range(4))
-    A = np.ones((3, 3)) / 9
-    mu = np.array(range(3))
-    sigma = np.array(range(1, 4))
-    pi = np.ones((3)) / 3
+    O = np.array(range(4), dtype=np.float_)
+    A = np.ones((3, 3), dtype=np.float_) / 9
+    mu = np.array(range(3), dtype=np.float_)
+    sigma = np.array(range(1, 4), dtype=np.float_)
+    pi = np.ones((3), dtype=np.float_) / 3
 
     def SMAPE(self, M1: npt.NDArray, M2: npt.NDArray):
         smape = (np.absolute(M1 - M2) / (np.absolute(M1) + np.absolute(M2)).mean()).mean() * 1e+2
